pyproject.toml
[build-system]
requires = ["uv_build>=0.9.10,<0.10.0"]
build-backend = "uv_build"

[project]
name = "workflows-acp"
version = "0.4.0-alpha"
description = "LlamaIndex Agent Workflows over the ACP wire"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agent-client-protocol>=0.7.1",
    "agentfs-sdk>=0.4.1",
    "anthropic>=0.75.0",
    "google-genai>=1.56.0",
    "llama-index-workflows>=2.11.6",
    "mcp-use>=1.5.1",
    "openai>=2.14.0",
    "psutil>=7.2.1",
    "pyyaml>=6.0.3",
    "typer>=0.21.0",
]

[project.scripts]
wfacp = "workflows_acp.cli:app"

[tool.uv.build-backend]
module-name = "workflows_acp"

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.10",
    "ty>=0.0.8",
]
